requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-qdrant
qdrant-client
python-dotenv